fix(python-wheel): refuse negative block dimensions at the cast, and read i16 magnitude in a wider type
Review follow-ups from CodeRabbit.

Two negative dimensions cancelled: `sample_count=-1 × channels=-1 × 4` is a
length a four-byte payload satisfies, so the block constructed and then failed
at `reshape` — the plausible-looking-wrong-answer path this cast exists to
close. Refused by name before the length check now. The Rust cast spells these
`u32`, where the case cannot arise.

`numpy.abs` keeps `int16`, where -32768 has no positive counterpart, so the
test fixture would report the wrong magnitude for an `i16` block carrying it.

+++ sdk/streamlib-python-wheel/tests/single_processor_under_test.py
@@ -82,7 +82,10 @@ class AudioBlockInspector:
                 "samples": block.interleaved_sample_bytes,
                 "shape": list(samples.shape),
                 "numpy_type": samples.dtype.str,
-                "loudest_sample": float(numpy.max(numpy.abs(samples))),
+                # Promoted before the magnitude: `numpy.abs` keeps `int16`,
+                # where -32768 has no positive counterpart and comes back
+                # negative.
+                "loudest_sample": float(numpy.max(numpy.abs(samples.astype("<f8")))),
                 "first_sample_timestamp_ns": block.first_sample_timestamp_ns,
                 "samples_are_a_view_over_the_bag_bytes": (
                     viewed.base is block.interleaved_sample_bytes
